deleteAppraiser working but no response from server
We can delete appraisers, but I am not getting a response from the server saying that it is deleted and refreshing the page by getting all appraisers.

If the next function in your subscribe method is not running, it's possible that the HTTP response from the server is not being returned as expected. Here are some things you can check:

Check the server logs: Check if the server is receiving the HTTP DELETE request and responding with a success status code (2xx). If the server is not sending a response or sending an error status code, the next function will not be called.

Check the network tab in your browser's developer tools: Open the network tab in your browser's developer tools and look for the HTTP DELETE request to the server. Check if the request is being sent and if a response is being received. If there is an error with the request or response, it will be indicated in the network tab.

Check the deleteAppraiser method in your api service: Check if the deleteAppraiser method in your api service is returning a valid HTTP response. It's possible that there is an issue with the implementation of the method that is causing the response to not be returned as expected.