we are able to assign orders to appraisers when creating a new order.
created service for getting all appraisers as an array
getting them as an array is important to use string interpolation for use in a mat-select element.
I do not understand how to do this otherwise as a MatTableDataSource. This is a DRY conflict. the getAllAppraisersArray is shared in the appraisersService, and could also be possibly implemented in the appraisers-view component.