Fix placeholder value instructing user on assigning orders
The problem: The user was unaware how to unassign an order or not assign one. Or at least not perfectly intuitive. The placeholder value now reads "Leave blank if not assigned", if the user goes to select an appraiser. If they don't click to select, then it will remain assigned or unassigned. This has a limitation that if the order is to be unassigned, and they do not hit the select dropdown, the order will remain assigned to that appraiser.