Error Fix: Catch duplicate Appraiser ID
The duplication of Appraiser ID's has been an issue that crashes the server.
The fix: by adding an if statement before calling the api service.
In addition a function was created to iterate through the array of existing appraisers , (an array that is updated when the dialogue to add an appraiser is created.), checking if the array contains an appraiser with the emp_id attempting to be used in the dialogue window.