Order Service implementing Check for Existing ID
FIX: Server crashed when sending an order with an existing ID, this has been changed to alert the user instead of sending to the DB.
Creation of orderService, implementing getting all Orders[].
Injecting this service into the assign-order dialogue.
Create function to check for an existing ID, return true or false.
Implement this check in the addOrder function before API call.