Check for existing ClientID
FIX: Server crashed when sending a new client with an existing ID, this has been changed to alert the user instead of sending to the DB.
Creation of clientService, implementing getting all Clients[].
Injecting this service into the add-client dialogue.
Create function to check for an existing ID, return true or false.
Implement this check in the addClient function before API call.